requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.2
bandit==1.6.2
boto3==1.13.2
botocore==1.16.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.42.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.0
django-embed-video==1.3.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.15.2
gitdb==4.0.4
GitPython==3.1.1
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pbr==5.4.5
Pillow==7.1.2
protobuf==3.6.1
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.14.0
smmap==3.0.2
social-auth-app-django==3.4.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==1.32.0
toml==0.10.1
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
